Improvement to performance for installing modules
By checking if the module exists before running the
the install command, runtime is cut down by about 2/3.

Note: Get-Module seems to take as long as attempting
to install the module.  However, Get-InstalledModule
is quick, and if the module is local, behaves no
differently than it would without the check.